0.1.2 Release - Bug fixes, plugin aliases, CommonJS, and more...
This is a big release with many improvements.  First haml-js is now a CommonJS module and is in the Tusk repository.  Thanks to Tom Robinson for helping with that.  Some of the plugins got aliases for people who didn't like the original name.  For example, you can now do `:javascript` instead of `:script` and `:for` instead of `:each`.  There were many bug fixes now that the code is starting to be actually used by myself and others.